+++ src/features/browse/extensions/Extensions.tsx
@@ -152,6 +152,11 @@ export function Extensions({ tabsMenuHeight }: { tabsMenuHeight: number }) {
         [groupedExtensions],
     );
 
+    const visibleExtensionNames = useMemo(
+        () => visibleExtensions.map((extension) => extension.name),
+        [visibleExtensions],
+    );
+
     const areReposDefined = !!extensionStoresRequest.data?.extensionStores.totalCount;
     const areMultipleReposInUse = useMemo(() => {
         if (!allExtensions.length) {
@@ -193,7 +198,7 @@ export function Extensions({ tabsMenuHeight }: { tabsMenuHeight: number }) {
 
     useAppAction(
         <>
-            <AppbarSearch />
+            <AppbarSearch suggestions={visibleExtensionNames} />
             <CustomTooltip title={t`Install external extension`}>
                 <IconButton
                     onClick={() => {